move project tool to built-in tools
