+++ src/vs/workbench/contrib/chat/common/tools/projectTreeTool.ts
@@ -1,0 +1,139 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (C) 2025 Posit Software, PBC. All rights reserved.
+ *  Licensed under the Elastic License 2.0. See LICENSE.txt for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import { CancellationToken } from '../../../../../base/common/cancellation.js';
+import { localize } from '../../../../../nls.js';
+import { IWorkspaceContextService } from '../../../../../platform/workspace/common/workspace.js';
+import { IExplorerService } from '../../../files/browser/files.js';
+import { CountTokensCallback, IToolData, IToolImpl, IToolInvocation, IToolResult } from '../languageModelToolsService.js';
+import { ExplorerItem } from '../../../files/common/explorerModel.js';
+import { SortOrder } from '../../../files/common/files.js';
+import { IToolInputProcessor } from './tools.js';
+
+/**
+ * Represents either a file (string) or a directory (tuple with string and children).
+ */
+export type DirectoryItem = string | [string, DirectoryItem[]];
+
+export const ExtensionProjectTreeId = 'positron_getProjectTree';
+export const InternalProjectTreeId = `${ExtensionProjectTreeId}_internal`;
+export const ProjectTreeToolData: IToolData = {
+	id: InternalProjectTreeId,
+	displayName: localize('chat.tools.getProjectTree', "Get Project Tree"),
+	modelDescription: `Construct a tree of the directories and files in the project.`,
+	tags: ['positron-assistant']
+};
+
+export class ProjectTreeTool implements IToolImpl {
+	constructor(
+		@IWorkspaceContextService private readonly _workspaceContextService: IWorkspaceContextService,
+		@IExplorerService private readonly _explorerService: IExplorerService,
+	) { }
+
+	async invoke(invocation: IToolInvocation, countTokens: CountTokensCallback, token: CancellationToken): Promise<IToolResult> {
+		const workspaceFolders = this._workspaceContextService.getWorkspace().folders;
+		if (workspaceFolders.length === 0) {
+			return {
+				content: [],
+				toolResultMessage: 'No workspace folders found.'
+			};
+		}
+
+		const workspaceTrees: DirectoryItem[][] = [];
+		const treeErrors: string[] = [];
+		const sortOrder = this._explorerService.sortOrderConfiguration.sortOrder;
+
+		for (const workspaceFolder of workspaceFolders) {
+			const explorerRoot = this._explorerService.findClosestRoot(workspaceFolder.uri);
+			if (!explorerRoot) {
+				treeErrors.push(`No explorer root found for ${workspaceFolder.name}`);
+				continue;
+			}
+			try {
+				const tree = await constructDirectoryTree(explorerRoot, sortOrder);
+				if (!tree) {
+					continue;
+				}
+				workspaceTrees.push(tree);
+			} catch (error) {
+				treeErrors.push(`Failed to generate tree for ${workspaceFolder.name}: ${error}`);
+			}
+		}
+
+		return {
+			content: workspaceTrees.map(dirTree => ({ kind: 'text', value: JSON.stringify(dirTree) })),
+			toolResultMessage: treeErrors.length > 0 ? treeErrors.join('\n') : undefined
+		};
+	}
+}
+
+export class ProjectTreeInputProcessor implements IToolInputProcessor {
+	processInput(input: any) {
+		// No input processing needed for this tool
+		return input;
+	}
+}
+
+const IGNORED_DIRECTORIES = [
+	'node_modules',
+	'__pycache__',
+	'dist',
+];
+const IGNORED_FILES = [
+	'.DS_Store',
+	'Thumbs.db',
+	'desktop.ini',
+	'.Rhistory',
+	'.RData',
+	'.Ruserdata',
+];
+const IGNORED_EXTENSIONS = [
+	'*.o',
+	'*.a',
+	'*.so',
+	'*.pyo',
+];
+
+const isIgnoredDirectory = (dirName: string): boolean => {
+	return dirName.startsWith('.') || IGNORED_DIRECTORIES.includes(dirName);
+};
+
+const isIgnoredFile = (fileName: string): boolean => {
+	return IGNORED_FILES.includes(fileName) || IGNORED_EXTENSIONS.some(ext => fileName.endsWith(ext));
+};
+
+/**
+ * Converts an explorer item to the directory item format
+ * @param item The explorer item to convert
+ * @returns A directory item array representing the file tree in the workspace
+ */
+export async function constructDirectoryTree(item: ExplorerItem, sortOrder: SortOrder): Promise<DirectoryItem[] | undefined> {
+	if (item.isDirectory) {
+		if (isIgnoredDirectory(item.name)) {
+			return undefined;
+		}
+
+		if (!item.hasChildren) {
+			return [[item.name, []]];
+		}
+
+		const children: DirectoryItem[] = [];
+		const itemChildren = await item.fetchChildren(sortOrder);
+		for (const child of itemChildren) {
+			const childEntries = await constructDirectoryTree(child, sortOrder);
+			if (childEntries === undefined) {
+				continue;
+			}
+			children.push(...childEntries);
+		}
+		return [[item.name, children]];
+	} else {
+		if (isIgnoredFile(item.name)) {
+			return undefined;
+		}
+		return [item.name];
+	}
+}
+
