+++ src/vs/workbench/contrib/positronAssistant/browser/positronAssistantService.ts
@@ -17,7 +17,6 @@ import { ILayoutService } from '../../../../platform/layout/browser/layoutServic
 import { showLanguageModelModalDialog } from './languageModelModalDialog.js';
 import { IConfigurationService } from '../../../../platform/configuration/common/configuration.js';
 import { Emitter } from '../../../../base/common/event.js';
-import { IPositronWorkspaceService } from '../../positronWorkspace/browser/positronWorkspaceService.js';
 
 /**
  * PositronAssistantService class.
@@ -42,7 +41,6 @@ export class PositronAssistantService extends Disposable implements IPositronAss
 		@IKeybindingService private readonly _keybindingService: IKeybindingService,
 		@ILayoutService private readonly _layoutService: ILayoutService,
 		@IConfigurationService private readonly _configurationService: IConfigurationService,
-		@IPositronWorkspaceService private readonly _positronWorkspaceService: IPositronWorkspaceService,
 	) {
 		super();
 	}
@@ -85,10 +83,6 @@ export class PositronAssistantService extends Disposable implements IPositronAss
 		const plot = this._plotService.positronPlotInstances.find(plot => plot.id === this._plotService.selectedPlotId);
 		const isPlotVisible = !!(plot instanceof PlotClientInstance && plot.lastRender);
 		return isPlotVisible ? plot.lastRender.uri : undefined;
-	}
-
-	async getProjectTree(): Promise<object> {
-		return await this._positronWorkspaceService.getProjectTree();
 	}
 
 	addLanguageModelConfig(source: IPositronLanguageModelSource): void {
